add preview monitor
